pyproject.toml
[project]
name = "obr"
version = "0.4.0"
description = "A tool to create and run OpenFOAM parameter studies"
authors = [
    {name = "Gregor Olenik", email = "go@hpsim.de"},
    {name = "Lukas Petermann", email = "lukas.petermann13@gmail.com"}
]
license = {text = "BSD-2-Clause"}
readme="Readme.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.6"
dependencies = [
    "click",
    "PyYAML",
    "signac==2.1.0",
    "signac-flow==0.26.1",
    "GitPython==3.1.31",
    "pandas",
    "DeepDiff",
    "jsonschema==4.19.1",
    "coloredlogs",
    "Owls @ git+https://github.com/greole/Owls.git@Owls2.0",
]

[project_urls]
Documentation= "https://obr.readthedocs.io/"
Changelog= "https://obr.readthedocs.io/en/latest/changelog.html"
Issues= "https://github.com/hpsim/obr/issues"

[project.scripts]
obr = "obr.cli:main"

[project.optional-dependencies]
doc = [
    "sphinx==6.2.1",
    "sphinx-autoapi==3.0.0",
    "sphinxcontrib.applehelp==1.0.7",
    "furo",
    "myst_parser"
]
test = [
    "pytest",
    "coverage",
    "gitpython"
]

[tool.black]
preview = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203"]

# [tool.setuptools]
